The image update
Added images of every single game in the collection. All of them are to be given file names that correspond to their ids in the database to show them on the correct page. If they are not already in the database, they are also added to it. This has currently been done to 70 out of 172 images.